+++ app/src/main/java/com/example/yandex_school_app/common/data/mapper/CategoryMapper.kt
@@ -1,6 +1,6 @@
 package com.example.yandex_school_app.common.data.mapper
 
-import com.example.yandex_school_app.category.domain.CategoryDomain
+import com.example.yandex_school_app.features.category.domain.CategoryDomain
 import com.example.yandex_school_app.common.data.CategoryNetwork
 
 class CategoryMapper {
