+++ app/src/main/java/com/example/yandex_school_app/common/domain/TransactionDomain.kt
@@ -1,6 +1,6 @@
 package com.example.yandex_school_app.common.domain
 
-import com.example.yandex_school_app.category.domain.CategoryDomain
+import com.example.yandex_school_app.features.category.domain.CategoryDomain
 
 data class TransactionDomain(
     val categoryDomain: CategoryDomain,
