+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/data/AccountStatistics.kt
@@ -1,4 +1,4 @@
-package com.example.yandex_school_app.cash_account.data
+package com.example.yandex_school_app.features.cash_account.data
 
 /**
  * Данная модель необходима для построения графика -> маппер для нее буду делать когда будет ясна структура графика
