+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/domain/AccountDomain.kt
@@ -1,4 +1,4 @@
-package com.example.yandex_school_app.cash_account.domain
+package com.example.yandex_school_app.features.cash_account.domain
 
 data class AccountDomain(
     val balance: String,
