+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/presentation/CashAccountScreen.kt
@@ -1,4 +1,4 @@
-package com.example.yandex_school_app.cash_account.presentation
+package com.example.yandex_school_app.features.cash_account.presentation
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -10,7 +10,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import com.example.yandex_school_app.R
-import com.example.yandex_school_app.cash_account.domain.AccountDomain
+import com.example.yandex_school_app.features.cash_account.domain.AccountDomain
 import com.example.yandex_school_app.common.domain.ListItemModelUI
 import com.example.yandex_school_app.common.presentation.ListItem
 import com.example.yandex_school_app.common.presentation.TypeListItem
