+++ app/src/main/java/com/example/yandex_school_app/features/category/domain/CategoryDomain.kt
@@ -1,4 +1,4 @@
-package com.example.yandex_school_app.category.domain
+package com.example.yandex_school_app.features.category.domain
 
 data class CategoryDomain(
     val name: String,
