+++ app/src/main/java/com/example/yandex_school_app/features/category/presentation/CategoryScreen.kt
@@ -1,4 +1,4 @@
-package com.example.yandex_school_app.category.presentation
+package com.example.yandex_school_app.features.category.presentation
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -18,7 +18,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.yandex_school_app.category.domain.CategoryDomain
+import com.example.yandex_school_app.features.category.domain.CategoryDomain
 import com.example.yandex_school_app.common.domain.ListItemModelUI
 import com.example.yandex_school_app.common.presentation.ListItem
 import com.example.yandex_school_app.common.presentation.TypeListItem
